// packages/ui/exchange-ui-core/src/utils/exchange/types.ts
export interface IProductDTO {
    deviceType?: string[];
    location?: string[];
    deviceVintage?: DeviceVintageDTO;
    generationFrom?: string;
    generationTo?: string;
    gridOperator?: string[];
}
// ...
export interface IAsset {
    id: string;
    address: string;
    tokenId: string;
    deviceId: string;
    generationFrom: string;
    generationTo: string;
}
// ...
export interface IOrder {
    id: string;
    side: OrderSide;
    validFrom: string;
    product: IProductDTO;
    price: number;
    startVolume: string;
    currentVolume: string;
    directBuyId: string;
    asset: IAsset;
    assetId: string;
    userId: string;
    filled?: number;
    demandId?: string;
}

export type Order = IOrder & { assetId: string; status: OrderStatus };
// ...
export const ANY_VALUE = 'Any';
export const ANY_OPERATOR = 'TH-ANY';